[#45] Rewrite to Multiple Public Libraries
Resolves #45
