+++ .travis.yml
@@ -17,13 +17,14 @@ matrix:
   - ghc: 8.2.2
   - ghc: 8.4.4
   - ghc: 8.6.5
+  - ghc: 8.8.1
 
 install:
   - cabal update
-  - cabal build all --enable-tests --enable-benchmarks
+  - cabal build --enable-tests --enable-benchmarks
 
 script:
-  - cabal test all --enable-tests
+  - cabal test --enable-tests
 
 notifications:
   email: false
