+++ CHANGELOG.md
@@ -1,0 +1,11 @@
+# Changelog
+
+`containers-backpack` uses [PVP Versioning][1].
+The changelog is available [on GitHub][2].
+
+## 0.0.0.0
+
+* Initially created.
+
+[1]: https://pvp.haskell.org
+[2]: https://github.com/kowainik/containers-backpack/releases
