--- cabal.project
@@ -1,11 +1,0 @@
-packages: containers-sig-readonly
-          containers-sig
-          containers-ordered-strict
-          containers-int-strict
-          containers-unordered-strict
-          containers-primitive
-          containers-contrib-readonly
-          containers-contrib
-          containers-example
-          containers-sig-laws
-          containers-sig-benchmarks
