+++ containers-backpack.cabal
@@ -1,0 +1,189 @@
+cabal-version:       3.0
+name:                containers-backpack
+version:             0.0.0.0
+synopsis:            Backpack signatures for Map-like interfaces
+description:         Backpack signatures for Map-like interfaces
+homepage:            https://github.com/kowainik/containers-backpack
+bug-reports:         https://github.com/kowainik/containers-backpack/issues
+license:             MPL-2.0
+license-file:        LICENSE
+author:              Dmitrii Kovanikov, Veronika Romashkina
+maintainer:          Kowainik <xrom.xkov@gmail.com>
+copyright:           2018-2019 Kowainik
+category:            Backpack, Data, Containers
+build-type:          Simple
+extra-doc-files:     CHANGELOG.md
+                     README.md
+tested-with:         GHC == 8.2.2
+                     GHC == 8.4.4
+                     GHC == 8.6.5
+                     GHC == 8.8.1
+
+source-repository head
+  type:                git
+  location:            https://github.com/kowainik/containers-backpack.git
+
+common common-options
+  build-depends:       base >= 4.10.1.0 && < 4.14
+
+  ghc-options:         -Wall
+                       -Wincomplete-uni-patterns
+                       -Wincomplete-record-updates
+                       -Wcompat
+                       -Widentities
+                       -Wredundant-constraints
+                       -Wmissing-export-lists
+                       -Wpartial-fields
+                       -fhide-source-paths
+  if impl(ghc >= 8.8.1)
+    ghc-options:       -Wmissing-deriving-strategies
+                       -Werror=missing-deriving-strategies
+
+  default-language:    Haskell2010
+  default-extensions:  ConstraintKinds
+                       DeriveGeneric
+                       DerivingStrategies
+                       InstanceSigs
+                       GeneralizedNewtypeDeriving
+                       LambdaCase
+                       OverloadedStrings
+                       RecordWildCards
+                       ScopedTypeVariables
+                       StandaloneDeriving
+                       TupleSections
+                       TypeApplications
+                       ViewPatterns
+
+-- signatures for containers
+library sig
+  import:              common-options
+  hs-source-dirs:      src/sig
+  signatures:          Map
+
+-- signature implementation for containers.Map.Strict
+library ordered-strict
+  import:              common-options
+  hs-source-dirs:      src/ordered-strict
+  exposed-modules:     Map.Ord
+  reexported-modules:  Map.Ord as Map
+  build-depends:       containers >= 0.5.10.1 && < 0.7
+  ghc-options:         -fexpose-all-unfoldings
+
+-- signature implementation for containers.IntMap.Strict
+library int-strict
+  import:              common-options
+  hs-source-dirs:      src/int-strict
+  exposed-modules:     Map.Int
+  reexported-modules:  Map.Int as Map
+  build-depends:       containers >= 0.5.10.1 && < 0.7
+                     , deepseq ^>= 1.4
+  ghc-options:         -fexpose-all-unfoldings
+
+-- signature implementation for unordered-containers.HashMap.Strict
+library unordered-strict
+  import:              common-options
+  hs-source-dirs:      src/unordered-strict
+  exposed-modules:     Map.Hash
+  reexported-modules:  Map.Hash as Map
+  build-depends:       hashable >= 1.2.7.0 && < 1.4
+                     , unordered-containers ^>= 0.2.7.0
+  ghc-options:         -fexpose-all-unfoldings
+
+-- signature implementation for primitive-containers.Map.Lifted
+library primitive
+  import:              common-options
+  hs-source-dirs:      src/primitive
+  exposed-modules:     Map.Prim
+  reexported-modules:  Map.Prim as Map
+  build-depends:       primitive-containers ^>= 0.4.0
+  ghc-options:         -fexpose-all-unfoldings
+
+-- implementations of various general functions via signatures
+library contrib
+  import:              common-options
+  hs-source-dirs:      src/contrib
+  exposed-modules:     Map.Contrib.Group
+  build-depends:       containers-backpack:sig
+
+-- property laws tests for readonly containers via signatures
+library laws
+  import:              common-options
+  hs-source-dirs:      src/laws
+  exposed-modules:     Map.Laws
+  build-depends:       containers-backpack:sig, QuickCheck ^>= 2.13
+
+-- benchmarks in terms of signatures
+library benchmarks
+  import:              common-options
+  hs-source-dirs:      src/benchmarks
+  exposed-modules:     Map.Bench
+  build-depends:       containers-backpack:sig, gauge ^>= 0.2.5, relude ^>= 0.5.0
+  -- TODO: ^ write in several lines when parsing of several lines with MPL is supported
+  -- * https://github.com/haskell/cabal/issues/5846
+
+executable example
+  import:              common-options
+  hs-source-dirs:      app
+  main-is:             Main.hs
+  build-depends:       contrib
+                     , int-strict
+                     , ordered-strict
+                     , unordered-strict
+
+  -- NOTE: need to omit prefix due to bug
+  -- * https://github.com/haskell/cabal/issues/6281
+  mixins:              contrib  (Map.Contrib.Group as Map.Contrib.Group.Int)
+                       requires (Map as Map.Int)
+                     , contrib  (Map.Contrib.Group as Map.Contrib.Group.Ord)
+                       requires (Map as Map.Ord)
+                     , contrib  (Map.Contrib.Group as Map.Contrib.Group.Hash)
+                       requires (Map as Map.Hash)
+
+test-suite laws-test
+  import:              common-options
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test
+  main-is:             Main.hs
+  build-depends:       hashable >= 1.2.7.0 && < 1.4
+                     , text
+                     , QuickCheck
+                     -- containers-backpack libraries
+                     , laws
+                     , int-strict
+                     , ordered-strict
+                     , unordered-strict
+
+  mixins:              laws     (Map.Laws as Map.Laws.Int)
+                       requires (Map as Map.Int)
+                     , laws     (Map.Laws as Map.Laws.Ord)
+                       requires (Map as Map.Ord)
+                     , laws     (Map.Laws as Map.Laws.Hash)
+                       requires (Map as Map.Hash)
+
+  ghc-options:         -threaded
+                       -rtsopts
+                       -with-rtsopts=-N
+
+benchmark simple-benchmark
+  import:              common-options
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      benchmark
+  main-is:             Main.hs
+
+  build-depends:       gauge ^>= 0.2.5
+                     -- containers-backpack libraries
+                     , benchmarks
+                     , int-strict
+                     , ordered-strict
+                     , unordered-strict
+
+  mixins:              benchmarks (Map.Bench as Map.Bench.Int)
+                       requires   (Map as Map.Int)
+                     , benchmarks (Map.Bench as Map.Bench.Ord)
+                       requires   (Map as Map.Ord)
+                     , benchmarks (Map.Bench as Map.Bench.Hash)
+                       requires   (Map as Map.Hash)
+
+  ghc-options:         -threaded
+                       -rtsopts
+                       -with-rtsopts=-N
