--- containers-contrib-readonly/containers-contrib.cabal
@@ -1,32 +1,0 @@
-name:                containers-contrib-readonly
-version:             0.0.0
-synopsis:            Different functions implemented via general containers-sig-readonly interface
-description:         Different functions implemented via general containers-sig-readonly interface
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-library
-  hs-source-dirs:      src
-  exposed-modules:     Map.Contrib.Laws
-                       Map.Contrib.Bench
-
-  build-depends:       base
-                     , QuickCheck
-                     , containers-sig-readonly
-                     , gauge
-                     , text
-                     , relude
-  ghc-options:         -Wall -fhide-source-paths
-  default-language:    Haskell2010
