--- containers-contrib/containers-contrib.cabal
@@ -1,32 +1,0 @@
-name:                containers-contrib
-version:             0.0.0
-synopsis:            Different functions implemented via general containers-sig interface
-description:         Different functions implemented via general containers-sig interface
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-license:             MPL-2.0
-license-file:        LICENSE
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-library
-  hs-source-dirs:      src
-  exposed-modules:     Map.Contrib.Group
-                       Map.Contrib.Laws
-
-  build-depends:       base
-                     , containers-sig
-                     , containers-sig-readonly
-                     , QuickCheck
-  ghc-options:         -Wall -fhide-source-paths
-  default-language:    Haskell2010
