--- containers-example/containers-example.cabal
@@ -1,44 +1,0 @@
-name:                containers-example
-version:             0.0.0
-synopsis:            Usage example of containers-sig
-description:         Usage example of containers-sig
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-license:             MPL-2.0
-license-file:        LICENSE
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-executable map-exe
-  hs-source-dirs:      app
-  main-is:             Main.hs
-  build-depends:       base
-                     , containers-int-strict
-                     , containers-ordered-strict
-                     , containers-unordered-strict
-                     , containers-primitive
-                     , containers-contrib-readonly
-                     , containers-contrib
-
-  mixins:              containers-contrib (Map.Contrib.Group as Map.Contrib.Group.Int)
-                                 requires (Map as Map.Int)
-                     , containers-contrib (Map.Contrib.Group as Map.Contrib.Group.Ord)
-                                 requires (Map as Map.Ord)
-                     , containers-contrib (Map.Contrib.Group as Map.Contrib.Group.Hash)
-                                 requires (Map as Map.Hash)
-
-                     , containers-contrib-readonly (Map.Contrib.Laws as Map.Contrib.Laws.Prim)
-                                          requires (Map as Map.Prim)
-
-  ghc-options:         -Wall -threaded -rtsopts -with-rtsopts=-N -fhide-source-paths
-  default-language:    Haskell2010
