--- containers-ordered-strict/CHANGELOG.md
@@ -1,12 +1,0 @@
-Change log
-==========
-
-`containers-ordered-strict` uses [PVP Versioning][1].
-The change log is available [on GitHub][2].
-
-0.0.0
-=====
-* Initially created.
-
-[1]: https://pvp.haskell.org
-[2]: https://github.com/kowainik/containers-backpack/releases
