--- containers-ordered-strict/containers-ordered-strict.cabal
@@ -1,31 +1,0 @@
-name:                containers-ordered-strict
-version:             0.0.0
-synopsis:            Backpack implementation of containers-sig for Data.Map.Strict
-description:         Backpack implementation of containers-sig for Data.Map.Strict
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-license:             MPL-2.0
-license-file:        LICENSE
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-library
-  hs-source-dirs:      src
-
-  exposed-modules:     Map.Ord
-  reexported-modules:  Map.Ord as Map
-
-  build-depends:       base, containers
-  ghc-options:         -Wall -fhide-source-paths
-                       -fexpose-all-unfoldings
-  default-language:    Haskell2010
