--- containers-primitive/containers-primitive.cabal
@@ -1,30 +1,0 @@
-name:                containers-primitive
-version:             0.0.0
-synopsis:            Backpack implementation of containers-sig for Data.Map.Lifted.Lifted
-description:         Backpack implementation of containers-sig for Data.Map.Lifted.Lifted
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-library
-  hs-source-dirs:      src
-
-  exposed-modules:     Map.Prim
-  reexported-modules:  Map.Prim as Map
-
-  build-depends:       base
-                     , primitive-containers ^>= 0.3.0
-  ghc-options:         -Wall -fhide-source-paths
-                       -fexpose-all-unfoldings
-  default-language:    Haskell2010
