--- containers-sig-laws/containers-sig-laws.cabal
@@ -1,52 +1,0 @@
-name:                containers-sig-laws
-version:             0.0.0
-synopsis:            Property tests for Map laws
-description:         Property tests for Map laws
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-license:             MPL-2.0
-license-file:        LICENSE
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-executable property-test-exe
-  hs-source-dirs:      app
-  main-is:             Main.hs
-  build-depends:       base
-                     , containers-int-strict
-                     , containers-ordered-strict
-                     , containers-unordered-strict
-                     , containers-primitive
-                     , containers-contrib
-                     , containers-contrib-readonly
-                     , text
-                     , QuickCheck
-                     , hashable
-
-  mixins:              containers-contrib-readonly (Map.Contrib.Laws as Map.Contrib.Laws.RO.Int)
-                                          requires (Map as Map.Int)
-                     , containers-contrib (Map.Contrib.Laws as Map.Contrib.Laws.Int)
-                                          requires (Map as Map.Int)
-                     , containers-contrib-readonly (Map.Contrib.Laws as Map.Contrib.Laws.RO.Ord)
-                                          requires (Map as Map.Ord)
-                     , containers-contrib (Map.Contrib.Laws as Map.Contrib.Laws.Ord)
-                                          requires (Map as Map.Ord)
-                     , containers-contrib-readonly (Map.Contrib.Laws as Map.Contrib.Laws.RO.Hash)
-                                          requires (Map as Map.Hash)
-                     , containers-contrib (Map.Contrib.Laws as Map.Contrib.Laws.Hash)
-                                          requires (Map as Map.Hash)
-                     , containers-contrib-readonly (Map.Contrib.Laws as Map.Contrib.Laws.RO.Prim)
-                                          requires (Map as Map.Prim)
-
-  ghc-options:         -Wall -threaded -rtsopts -with-rtsopts=-N -fhide-source-paths
-  default-language:    Haskell2010
