--- containers-sig-readonly/containers-sig-readonly.cabal
@@ -1,27 +1,0 @@
-name:                containers-sig-readonly
-version:             0.0.0
-synopsis:            Backpack signatures for readonly Map-like interfaces
-description:         Backpack signatures for readonly Map-like interfaces
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-license:             MPL-2.0
-license-file:        LICENSE
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-library
-  hs-source-dirs:      src
-  signatures:          Map
-  build-depends:       base
-  ghc-options:         -Wall -fhide-source-paths
-  default-language:    Haskell2010
