--- containers-unordered-strict/containers-unordered-strict.cabal
@@ -1,33 +1,0 @@
-name:                containers-unordered-strict
-version:             0.0.0
-synopsis:            Backpack implementation of containers-backpack for Data.HashMap.Strict
-description:         Backpack implementation of containers-backpack for Data.HashMap.Strict
-homepage:            https://github.com/kowainik/containers-backpack
-bug-reports:         https://github.com/kowainik/containers-backpack/issues
-license:             MPL-2.0
-license-file:        LICENSE
-author:              Kowainik
-maintainer:          xrom.xkov@gmail.com
-copyright:           2018 Kowainik
-category:            Backpack, Data, Containers
-build-type:          Simple
-extra-doc-files:     CHANGELOG.md
-cabal-version:       2.0
-tested-with:         GHC == 8.4.3
-
-source-repository head
-  type:                git
-  location:            https://github.com/kowainik/containers-backpack.git
-
-library
-  hs-source-dirs:      src
-
-  exposed-modules:     Map.Hash
-  reexported-modules:  Map.Hash as Map
-
-  build-depends:       base
-                     , hashable
-                     , unordered-containers
-  ghc-options:         -Wall -fhide-source-paths
-                       -fexpose-all-unfoldings
-  default-language:    Haskell2010
